+++ src/guitext.h
@@ -97,7 +97,7 @@ enum GuiStrings {
 extern char *gui_strings[STRINGS_MAX];
 extern char *gui_strings_data;
 extern char *gui_strings_data_end;
-extern ulong text_buf_pos;
+extern u32 text_buf_pos;
 
 void snprint_dh_time_duration(char *out, ulong outlen, long ndays, short nhours);
